requirements.txt
streamlit
pandas
notion-client
google-generativeai
scikit-learn
streamlit-oauth